refactor: defer loading data when required
